Neu: Zaubersprüche können in LUA programmiert werden, und per XML-Datei (siehe spells.xml) in ein Spiel integriert werden. Ich habe angefangen, die Item-erzeugenden Zauber aus dem Code in Skripte zu verlagern.

+++ src/common/kernel/eressea.h
@@ -51,7 +51,7 @@ typedef short luxury_t;
 typedef short weapon_t;
 typedef short item_t;
 typedef short resource_t;
-typedef short spellid_t;
+typedef unsigned int spellid_t;
 
 struct plane;
 struct order;
